requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3

# Git analysis
GitPython==3.1.41

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP client for Ollama
httpx==0.26.0
aiohttp==3.9.3

# Data processing
pydantic==2.5.3
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
